+++ src/main.ts
@@ -3,6 +3,7 @@ import App from "./App.vue";
 import router from "./router";
 import { i18n } from "./i18n";
 
+import "element-plus/theme-chalk/dark/css-vars.css";
 import "./assets/main.css";
 
 const app = createApp(App);